Refactor ClassAnalyzer into smaller handlers
