+++ tests/codegen/handlers/test_attribute_merge.py
@@ -1,0 +1,63 @@
+from tests.factories import AttrFactory
+from tests.factories import ClassFactory
+from tests.factories import FactoryTestCase
+from xsdata.codegen.handlers import AttributeMergeClassHandler
+from xsdata.models.codegen import Restrictions
+
+
+class AttributeMergeHandlerTests(FactoryTestCase):
+    def setUp(self):
+        super().setUp()
+        self.processor = AttributeMergeClassHandler
+
+    def test_process(self):
+        one = AttrFactory.attribute(fixed=True)
+        one_clone = one.clone()
+        restrictions = Restrictions(min_occurs=10, max_occurs=15)
+        two = AttrFactory.element(restrictions=restrictions, fixed=True)
+        two_clone = two.clone()
+        two_clone.restrictions.min_occurs = 5
+        two_clone.restrictions.max_occurs = 5
+        two_clone_two = two.clone()
+        two_clone_two.restrictions.min_occurs = 4
+        two_clone_two.restrictions.max_occurs = 4
+        three = AttrFactory.element()
+        four = AttrFactory.enumeration()
+        four_clone = four.clone()
+        five = AttrFactory.element()
+        five_clone = five.clone()
+        five_clone_two = five.clone()
+
+        target = ClassFactory.create(
+            attrs=[
+                one,
+                one_clone,
+                two,
+                two_clone,
+                two_clone_two,
+                three,
+                four,
+                four_clone,
+                five,
+                five_clone,
+                five_clone_two,
+            ]
+        )
+
+        winners = [one, two, three, four, five]
+
+        self.processor.process(target)
+        self.assertEqual(winners, target.attrs)
+
+        self.assertTrue(one.fixed)
+        self.assertIsNone(one.restrictions.min_occurs)
+        self.assertIsNone(one.restrictions.max_occurs)
+        self.assertFalse(two.fixed)
+        self.assertEqual(4, two.restrictions.min_occurs)
+        self.assertEqual(24, two.restrictions.max_occurs)
+        self.assertIsNone(three.restrictions.min_occurs)
+        self.assertIsNone(three.restrictions.max_occurs)
+        self.assertIsNone(four.restrictions.min_occurs)
+        self.assertIsNone(four.restrictions.max_occurs)
+        self.assertEqual(0, five.restrictions.min_occurs)
+        self.assertEqual(3, five.restrictions.max_occurs)
